add new module for search related routes
