+++ server/main.py
@@ -42,7 +42,6 @@ app.jinja_env.globals['NAME'] = app.config['NAME']
 app.jinja_env.globals['BASE_HTML'] = (
     'sustainability/base.html' if app.config['SUSTAINABILITY'] else 'base.html')
 
-_MAX_SEARCH_RESULTS = 1000
 
 WARM_UP_ENDPOINTS = [
     "/api/choropleth/geojson?placeDcid=country/USA&placeType=County",
@@ -82,52 +81,9 @@ def translator_handler():
                                  sample_query=translator.SAMPLE_QUERY)
 
 
-# TODO(beets): Move this to a separate handler so it won't be installed on all apps.
-@app.route('/search')
-def search():
-    return flask.render_template('search.html')
-
-
 @app.route('/healthz')
 def healthz():
     return "very healthy"
-
-
-# TODO(beets): Move this to a separate handler so it won't be installed on all apps.
-@app.route('/search_dc')
-def search_dc():
-    """Add DC API powered search for non-place searches temporarily"""
-    query_text = request.args.get('q', '')
-    max_results = int(request.args.get('l', _MAX_SEARCH_RESULTS))
-    search_response = dc.search(query_text, max_results)
-
-    # Convert from search results to template dictionary.
-    results = []
-    query_tokens = set(query_text.lower().split())
-    for section in search_response.get('section', []):
-        entities = []
-        for search_entity in section['entity']:
-            entity = {}
-            entity['name'] = search_entity['name']
-            entity['dcid'] = search_entity['dcid']
-            name_tokens = search_entity['name'].lower().split()
-            for i, t in enumerate(name_tokens):
-                name_tokens[i] = t.strip("'")
-            name_tokens = set(name_tokens)
-            if not name_tokens & query_tokens:
-                continue
-            entity['rank'] = len(name_tokens & query_tokens) / len(name_tokens |
-                                                                   query_tokens)
-            entities.append(entity)
-        entities = sorted(entities, key=lambda e: (e['rank']), reverse=True)
-        if entities:
-            results.append({
-                'type': section['typeName'],
-                'entities': entities,
-            })
-    return flask.render_template('search_dc.html',
-                                 query_text=query_text,
-                                 results=results)
 
 
 # TODO(beets): Move this to a separate handler so it won't be installed on all apps.
