+++ server/templates/search_landing.html
@@ -1,0 +1,52 @@
+{#
+ Copyright 2022 Google LLC
+
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+#}
+{% extends 'base.html' %}
+
+{% set main_id = 'search' %}
+{% set title = 'Search' %}
+{% set subpage_title = _('Search') %}
+{% set subpage_url = url_for('search.search_landing') %}
+
+{% block head %}
+<link rel="stylesheet" href={{url_for('static', filename='css/search.min.css', t=config['VERSION'])}} >
+{% endblock %}
+
+{% block content %}
+<div class="container">
+  <article class="mb-5 mt-5">
+    <h1 class="mb-3">Search Data Commons content</h1>
+    <div class="gcse-searchbox-only" data-resultsUrl="/search"></div>
+  </article>
+
+  <article>
+    <h1 class="mb-3">Search Places in Data Commons</h1>
+    <div class="search border mb-4">
+      <div id="location-field">
+        <div id="search-icon"></div>
+        <input id="place-autocomplete"
+          placeholder="{% trans %}Enter a country, state, county or city{% endtrans %}" type="text" />
+      </div>
+    </div>
+  </article>
+</div>
+
+<script
+  src={{url_for('static', filename='search_landing.js', t=config['VERSION'])}}></script>
+<script async src="https://cse.google.com/cse.js?cx=010079880385165835740%3Aosnv3q-sw08"></script>
+<script
+  src="https://maps.googleapis.com/maps/api/js?key={{maps_api_key}}&libraries=places&language={{locale}}"
+  async defer></script>
+{% endblock %}
