+++ static/css/search.scss
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 @import 'base';
+@import 'place/place_search';
 
 .gcsc-branding a,
 .gcsc-more-maybe-branding-root a {
